refactor: initial structure update
