--- client/.gitignore
@@ -1,2 +1,0 @@
-/target
-Cargo.lock
