+++ client/src/lib.rs
@@ -1,6 +1,35 @@
+//! # Eigen Trust
+//!
+//! A library for managing trust in a distributed network with zero-knowledge
+//! features.
+//!
+//! ## Main characteristics:
+//! **Self-policing** - the shared ethics of the user population is defined and
+//! enforced by the peers themselves and not by some central authority.
+//!
+//! **Minimal** - computation, infrastructure, storage, and message complexity
+//! are reduced to a minimum.
+//!
+//! **Incorruptible** - Reputation should be obtained by consistent good
+//! behavior through several transactions. This is enforced for all users, so no
+//! one can cheat the system and obtain a higher reputation. It is also
+//! resistant to malicious collectives.
+//!
+//! ## Implementation
+//! The library is implemented according to the original [Eigen Trust paper](http://ilpubs.stanford.edu:8090/562/1/2002-56.pdf).
+//! It is developed under the Ethereum Foundation grant.
+
 pub mod att_station;
+pub mod epoch;
+pub mod error;
+pub mod ethereum;
+pub mod manager;
 pub mod utils;
 
+use crate::manager::{
+	attestation::{Attestation, AttestationData},
+	NUM_NEIGHBOURS,
+};
 use att_station::{AttestationData as AttData, AttestationStation as AttStation};
 use eigen_trust_circuit::{
 	calculate_message_hash,
@@ -8,10 +37,6 @@ use eigen_trust_circuit::{
 	halo2::halo2curves::{bn256::Fr as Scalar, FieldExt},
 	utils::to_short,
 	ProofRaw,
-};
-use eigen_trust_server::manager::{
-	attestation::{Attestation, AttestationData},
-	NUM_NEIGHBOURS,
 };
 use ethers::{
 	abi::Address,
@@ -152,6 +177,7 @@ impl EigenTrustClient {
 #[cfg(test)]
 mod test {
 	use crate::{
+		manager::NUM_NEIGHBOURS,
 		utils::{deploy_as, deploy_et_wrapper, deploy_verifier},
 		ClientConfig, EigenTrustClient,
 	};
@@ -159,7 +185,6 @@ mod test {
 		utils::{read_bytes_data, read_json_data},
 		ProofRaw,
 	};
-	use eigen_trust_server::manager::NUM_NEIGHBOURS;
 	use ethers::{abi::Address, utils::Anvil};
 
 	#[tokio::test]
