--- server/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "eigen-trust-server"
-version = "0.2.0"
-description = "A library for managing trust in a distributed network with zero-knowledge features."
-repository = "https://github.com/lazovicff/eigen-trust"
-license = "MIT"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ethers = { version = "=1.0.2", features = ["ethers-solc", "ws"] }
-ethers-contract = "=1.0.0"
-bs58 = "0.4.0"
-rand = "0.8"
-futures = "0.3.21"
-tokio = { version = "1.18", features = ["time", "macros", "rt-multi-thread", "net"] }
-async-trait = "0.1"
-log = "0.4"
-hyper = { version = "0.14", features = ["server", "http1", "http2", "tcp"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-once_cell = "1.15.0"
-
-eigen-trust-circuit = { path = "../circuit" }
-
-[lib]
-doctest = false
